Move workload implementations to separate classes
