+++ csb/src/main/scala/edu/msstate/dasi/csb/Benchmark.scala
@@ -158,7 +158,7 @@ object Benchmark {
   }
 
   /**
-   * Runs a set of workloads on a specified graph and outputs the results.
+   * Runs a set of workloads on a specified graph and outputs the execution time.
    *
    * @param config  Group of configuration variables, adjusted by the command-line inputs.
    * @param factory Spark ComponentFactory object for the application.
@@ -170,90 +170,8 @@ object Benchmark {
 
     val graph = Util.time( "Load graph", loader.loadGraph(config.graphPrefix, config.partitions) )
 
-    val workloadEngine = factory.getWorkloadEngine
-    val workloads = config.workloads
-
-    val srcWorkloads = Seq("bfs", "closeness-centrality", "sssp")
-
-    val all = workloads.contains("all")
-
-    if ( all || workloads.contains("count-vertices") ) Util.time("Count vertices", workloadEngine.countVertices(graph))
-    if ( all || workloads.contains("count-edges") ) Util.time("Count edges", workloadEngine.countEdges(graph))
-
-    if ( all || workloads.contains("degree") ) Util.time("Degree", workloadEngine.degree(graph))
-    if ( all || workloads.contains("in-degree") ) Util.time("In-degree", workloadEngine.inDegree(graph))
-    if ( all || workloads.contains("out-degree") ) Util.time("Out-degree", workloadEngine.outDegree(graph))
-
-    if ( all || workloads.contains("pagerank") ) {
-      Util.time("PageRank", workloadEngine.pageRank(graph))
-    }
-
-    if ( all || workloads.contains("neighbors") ) Util.time("Neighbors", workloadEngine.neighbors(graph))
-    if ( all || workloads.contains("in-neighbors") ) Util.time("In-neighbors", workloadEngine.inNeighbors(graph))
-    if ( all || workloads.contains("out-neighbors") ) Util.time("Out-neighbors", workloadEngine.outNeighbors(graph))
-
-    if ( all || workloads.contains("in-edges") ) Util.time("Degree", workloadEngine.inEdges(graph))
-    if ( all || workloads.contains("out-edges") ) Util.time("In-degree", workloadEngine.outEdges(graph))
-
-    if ( all || workloads.contains("connected-components") ) {
-      Util.time("Connected Components", workloadEngine.connectedComponents(graph))
-    }
-
-    if ( all || workloads.contains("triangle-count") ) {
-      Util.time("Triangle Counting", workloadEngine.triangleCount(graph))
-    }
-
-    if ( all || workloads.contains("strongly-connected-components") ) {
-      Util.time("Strongly Connected Components", workloadEngine.stronglyConnectedComponents(graph, config.iterations))
-    }
-
-    if ( all || workloads.contains("betweenness-centrality") ) {
-      Util.time("Betweenness Centrality", workloadEngine.betweennessCentrality(graph, config.iterations))
-    }
-
-    if ( all || workloads.intersect(srcWorkloads).nonEmpty ) {
-      var srcVertex = config.srcVertex
-
-      if (srcVertex == 0L) {
-        println("Source not specified, extracting random ID from the input graph...")
-
-        srcVertex = graph.vertices.zipWithIndex()
-          .map{ case ((id, _), index) => (index, id) }
-          .lookup( (Random.nextDouble * graph.numVertices).toLong)
-          .head
-
-        println(s"Source vertex: $srcVertex")
-      }
-
-      if ( all || workloads.contains("bfs") ) {
-        var dstVertex = config.dstVertex
-
-        if (dstVertex == 0L) {
-          println("Destination not specified, extracting random ID from the input graph...")
-
-          dstVertex = graph.vertices.zipWithIndex()
-            .map{ case ((id, _), index) => (index, id) }
-            .lookup( (Random.nextDouble * graph.numVertices).toLong)
-            .head
-
-          println(s"Destination vertex: $dstVertex")
-        }
-
-        Util.time("Breadth-first Search", workloadEngine.bfs(graph, srcVertex, dstVertex))
-      }
-
-      if ( all || workloads.contains("closeness-centrality") ) {
-        Util.time("Closeness Centrality", workloadEngine.closenessCentrality(graph, srcVertex))
-      }
-
-      if ( all || workloads.contains("sssp") ) {
-        Util.time("Single Source Shortest Path", workloadEngine.sssp(graph, srcVertex))
-      }
-    }
-
-    if ( all || workloads.contains("subgraph-isomorphism") ) {
-      val pattern = Util.time("Load pattern", loader.loadGraph(config.patternPrefix, config.partitions))
-      Util.time("Subgraph Isomorphism", workloadEngine.subgraphIsomorphism(graph, pattern))
+    for (workload <- factory.getWorkloads) {
+      Util.time(s"${workload.name} workload", workload.run(graph))
     }
 
     true
